Support to render a GitHub users to be a link
