+++ main_test.go
@@ -2,9 +2,6 @@ package main
 
 import (
 	"bytes"
-	"github.com/h2non/gock"
-	"io/ioutil"
-	"net/http"
 	"testing"
 )
 import "github.com/stretchr/testify/assert"
@@ -119,42 +116,6 @@ content
 	}
 }
 
-func Test_printContributor(t *testing.T) {
-	type args struct {
-		owner string
-		repo  string
-	}
-	tests := []struct {
-		name       string
-		args       args
-		prepare    func()
-		wantOutput func() string
-	}{{
-		name: "normal case",
-		args: args{
-			owner: "linuxsuren",
-			repo:  "yaml-readme",
-		},
-		prepare: func() {
-			gock.New("https://api.github.com").
-				Get("/repos/linuxsuren/yaml-readme/contributors").
-				Reply(http.StatusOK).
-				File("data/yaml-readme.json")
-		},
-		wantOutput: func() string {
-			data, _ := ioutil.ReadFile("data/yaml-readme-contributors.txt")
-			return string(data)
-		},
-	}}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			defer gock.Off()
-			tt.prepare()
-			assert.Equalf(t, tt.wantOutput(), printContributors(tt.args.owner, tt.args.repo), "printContributors(%v, %v)", tt.args.owner, tt.args.repo)
-		})
-	}
-}
-
 func Test_printStarHistory(t *testing.T) {
 	type args struct {
 		owner string
@@ -216,10 +177,10 @@ func Test_dataRender(t *testing.T) {
 		want: ":x:",
 	}, {
 		name: "normal string value fake",
-		args: args {
-			data:"fake",
+		args: args{
+			data: "fake",
 		},
-		want:"fake",
+		want: "fake",
 	}}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
